fix strlen type

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: anachat <anachat@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/22 20:11:18 by anachat           #+#    #+#             */
-/*   Updated: 2024/10/29 15:52:29 by anachat          ###   ########.fr       */
+/*   Updated: 2024/11/09 16:38:46 by anachat          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
